block copy if file not found

+++ Tools/compile.sh
@@ -5,4 +5,6 @@ cd $DIR/..
 if [ ! -e liblinphone-sdk ] ; then
 exec make -j 8
 fi
+if [ -e libopenh264.a  ]; then
 cp -f libopenh264.a liblinphone-sdk/apple-darwin/lib/libopenh264.a
+fi
